Added two starting files
The two files I have added are mandelbrot.cpp and settings.txt. mandelbrot.cpp is just full of comments because I am just setting it up to work on. settings.txt will be filled in later, now it just says what it is and the purpose of it being used.